+++ include/private/configurationreader.hpp
@@ -26,6 +26,7 @@
 #define DB_PORT_ENV_VAR_NAME "DBAAS_SERVICE_PORT"
 #define SENTINEL_PORT_ENV_VAR_NAME "DBAAS_SERVICE_SENTINEL_PORT"
 #define SENTINEL_MASTER_NAME_ENV_VAR_NAME "DBAAS_MASTER_NAME"
+#define DB_CLUSTER_ADDR_LIST_ENV_VAR_NAME "DBAAS_CLUSTER_ADDR_LIST"
 
 #include <iosfwd>
 #include <string>
@@ -72,6 +73,8 @@ namespace shareddatalayer
         std::string sentinelPortEnvVariableValue;
         const std::string sentinelMasterNameEnvVariableName;
         std::string sentinelMasterNameEnvVariableValue;
+        const std::string dbClusterAddrListEnvVariableName;
+        std::string dbClusterAddrListEnvVariableValue;
         boost::optional<boost::property_tree::ptree> jsonDatabaseConfiguration;
         std::string sourceForDatabaseConfiguration;
         std::unordered_map<std::string, std::pair<boost::property_tree::ptree, std::string>> jsonNamespaceConfigurations;